Module Jupyter_kernel.Log

val src : Logs.Src.t

Logs source for jupyter-kernel

val file_reporter : string -> Logs.reporter

Logs reporter that writes into the given file.

val open_log_file : string -> unit

Open log file using Logs. This will add a reporter to Logs, and is a bit brittle as it might be replaced. See file_reporter to build the underlying reporter and install it yourself.

val log : string -> unit

val logf : ('a, unit, string, unit) Stdlib.format4 -> 'a

Logs for the src above.

include Logs.LOG

val msg : Logs.level -> 'a Logs.log

val app : 'a Logs.log

val err : 'a Logs.log

val warn : 'a Logs.log

val info : 'a Logs.log

val debug : 'a Logs.log

val kmsg : (unit -> 'b) -> Logs.level -> ('a, 'b) Logs.msgf -> 'b

val on_error : ?level:Logs.level -> ?header:string -> ?tags:Logs.Tag.set -> pp:(Stdlib.Format.formatter -> 'b -> unit) -> use:('b -> 'a) -> ('a, 'b) Stdlib.result -> 'a

val on_error_msg : ?level:Logs.level -> ?header:string -> ?tags:Logs.Tag.set -> use:(unit -> 'a) -> ('a, [ `Msg of string ]) Stdlib.result -> 'a