Module Jupyter_kernel__Message

type content =

| Connect_request

| Comm_info_request of Jupyter_kernel.Protocol_t.comm_info_request

| Kernel_info_request

| Shutdown_request of Jupyter_kernel.Protocol_t.shutdown

| Execute_request of Jupyter_kernel.Protocol_t.execute_request

| Inspect_request of Jupyter_kernel.Protocol_t.inspect_request

| Complete_request of Jupyter_kernel.Protocol_t.complete_request

| Is_complete_request of Jupyter_kernel.Protocol_t.is_complete_request

| History_request of Jupyter_kernel.Protocol_t.history_request

| Connect_reply of Jupyter_kernel.Protocol_t.connect_reply

| Comm_info_reply

| Kernel_info_reply of Jupyter_kernel.Protocol_t.kernel_info_reply

| Shutdown_reply of Jupyter_kernel.Protocol_t.shutdown

| Execute_reply of Jupyter_kernel.Protocol_t.execute_reply

| Inspect_reply of Jupyter_kernel.Protocol_t.inspect_reply

| Complete_reply of Jupyter_kernel.Protocol_t.complete_reply

| Is_complete_reply of Jupyter_kernel.Protocol_t.is_complete_reply

| History_reply of Jupyter_kernel.Protocol_t.history_reply

| Status of Jupyter_kernel.Protocol_t.status

| Execute_input of Jupyter_kernel.Protocol_t.pyin

| Execute_result of Jupyter_kernel.Protocol_t.pyout

| Execute_error of Jupyter_kernel.Protocol_t.execute_error

| Stream of Jupyter_kernel.Protocol_t.stream

| Clear of Jupyter_kernel.Protocol_t.clear_output

| Display_data of Jupyter_kernel.Protocol_t.display_data

| Comm_open

val content_of_json : Jupyter_kernel.Protocol_t.header_info -> string -> content

val json_of_content : content -> string

val msg_type_of_content : content -> string

type t = private {

ids : string array;

hmac : string;

header : Jupyter_kernel.Protocol_t.header_info;

parent : Jupyter_kernel.Protocol_t.header_info;

meta : string;

content : content;

raw : string array;

}

val log : string -> t -> unit

val recv : [ `Router ] Lwt_zmq.Socket.t -> t Lwt.t

val make : parent:t -> msg_type:string -> content -> t

make a message with the given type and content, copying header from parent

val make_first : msg_type:string -> content -> t

make a message with the given type and content, with a fresh header

val send : ?⁠key:string -> [< `Router | `Pub ] Lwt_zmq.Socket.t -> t -> unit Lwt.t

send sock msg sends the message.

parameter key

the shared key for signing messages

val wrap_retry : ('a -> 'b Lwt.t) -> 'a -> 'b Lwt.t

Wrapper to retry upon EAGAIN